Streamlined some code
I removed the for loop around the functions

I commented out all if!XthingExists code, so that my checklist doesn't get anymore spam

I changed functionOutput to a let, so that it can change for each if statement.

I replaced the original function code with the variable to prevent the errors.