Config update progress
You can now configure the names of both tasks and the daily's items.

I started progress towards the next major change, that being automatic detection of which curses challenges your in and setting up the checklist item appropriately

I removed comments on some of the config stuff since people won't be changing it in the main file.

I also updated the to-do list to reflect my progress